# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hostack API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/copilot/chat:
    post:
      operationId: copilotChat
      tags: [copilot]
      summary: Stream AI copilot chat response for a project
      description: Accepts a user message and optional conversation history, returns an SSE stream of AI-generated tokens.
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopilotChatBody"
      responses:
        "200":
          description: SSE stream of chat tokens
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /projects/{projectId}/domains:
    get:
      operationId: listProjectDomains
      tags: [domains]
      summary: List custom domains for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of custom domains
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomDomainListResponse"
        "401":
          description: Unauthorized
        "404":
          description: Project not found
    post:
      operationId: addProjectDomain
      tags: [domains]
      summary: Add a custom domain to a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDomainBody"
      responses:
        "201":
          description: Domain added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomDomainResponse"
        "400":
          description: Invalid domain
        "401":
          description: Unauthorized
        "409":
          description: Domain already exists
  /projects/{projectId}/domains/{domainId}:
    delete:
      operationId: removeProjectDomain
      tags: [domains]
      summary: Remove a custom domain from a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: domainId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Domain removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
        "401":
          description: Unauthorized
        "404":
          description: Domain not found
  /projects/{projectId}/domains/{domainId}/verify:
    post:
      operationId: verifyProjectDomain
      tags: [domains]
      summary: Verify DNS configuration for a custom domain
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: domainId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomDomainResponse"
        "401":
          description: Unauthorized
        "404":
          description: Domain not found
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CustomDomain:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        domain:
          type: string
        status:
          type: string
          enum: [pending, active, failed]
        verificationToken:
          type: string
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - domain
        - status
        - verificationToken
        - createdAt
    CustomDomainListResponse:
      type: object
      properties:
        domains:
          type: array
          items:
            $ref: "#/components/schemas/CustomDomain"
      required:
        - domains
    CustomDomainResponse:
      type: object
      properties:
        domain:
          $ref: "#/components/schemas/CustomDomain"
      required:
        - domain
    AddDomainBody:
      type: object
      properties:
        domain:
          type: string
      required:
        - domain
...
    CopilotChatBody:
      type: object
      properties:
        message:
          type: string
        history:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
            required:
              - role
              - content
      required:
        - message