# lib/api-zod/src/generated/types/customDomain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hostack API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CustomDomainStatus } from "./customDomainStatus";

export interface CustomDomain {
  id: string;
  projectId: string;
  domain: string;
  status: CustomDomainStatus;
  verificationToken: string;
  verifiedAt?: Date | null;
  createdAt: Date;
}